implements toast notifications on song add
