+++ src/App.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-
+import { ToastProvider } from 'react-toast-notifications'
 import Router from 'Router'
 import { ThemeProvider } from 'emotion-theming'
 import { Global } from '@emotion/core'
@@ -12,17 +12,19 @@ const App: React.FC = () => {
   return (
     <AuthContextProvider>
       <ThemeProvider theme={theme}>
-        <Global
-          styles={css({
-            body: {
-              color: 'text',
-              bg: 'background',
-              lineHeight: '1.5',
-              m: 0,
-            },
-          })}
-        />
-        <Router />
+        <ToastProvider>
+          <Global
+            styles={css({
+              body: {
+                color: 'text',
+                bg: 'background',
+                lineHeight: '1.5',
+                m: 0,
+              },
+            })}
+          />
+          <Router />
+        </ToastProvider>
       </ThemeProvider>
     </AuthContextProvider>
   )
